Update content rendering to match FN1
To make it easy to create websites and applications with a consistent look, Factory Number One now supports a `.contentBlock`, which contains a `.userContent` div. Separating the layout container from the content container makes it possible to add other structure information next to the main body content. Breaking the taxonomy down to generic components will make it easy to replicate the same look and feel across other sites and applications.