Cover Photo: check for actual image file
It was possible to "Add new media" for a cover photo without
actually uploading an image file.  I have added checks to
make sure we actually have an image file when rendering the
page.

Also, the image file field on media has been made a requied
field for the media bundle.