potential fix: info card text overflow
note:

This change nukes the pre-determined height set for every card. Instead, this uses Flex's ability to get the height of the greatest in the row and then sets the children to match.

So, what happens when one card in the row is one sentence in length and the card next to it has five sentences, is it problematic to have such a spacing discrepancy?

Is it better to at least see the text in full, without a hidden overflow, despite having different but similar sized info cards? Or, is it better to truncate text and make these all exactly uniform?

I'm open to ideas and conversation, outlining examples of cases where we might see such. Perhaps even tackling both solutions and doing user testing to see how folks feel.

ref #148 / #12