Replaced drupal search form in header
The drupal search form was provided by the database search module. The Solr search api does not
provide a form block; so we have to write our own form for site header.  I've tried to simplify the
form markup, while keeping most of the ID and classes the same.

Updates #214