Board meeting data now organized by event_id
We don't have the time information in the OnBoard database.  However, we can use the start time of the Google Calendar events, provided we have one.

OnBoard now organizes the meeting files by date and by event_id.  Meetings that are not in a Google Calendar will have an event_id of 0.  This means if there are multiple meetings on the same day, and they're not in Google Calendar, then they'll still be lumped together.

For meetings that are in Google Calendar, we can read the even information and display the start time from the calendar.

Updates #237
Updates #220