CSS: Info cards must be in an info-list
There is not enough markup to select info-cards to flex when they are not in an info-list.  This removes the CSS that attempted to do so.  Existing content on the site that has bare info-cards will need to be updated to preserve the info-card layout.  I think this currently only applies to Department pages.

Updates #211