Fixed: extra spacing in Drupal 11
The theme CSS expects to have FN1 declared first in the load order, so that theme styles can override the FN1 styles.

In Drupal 10, the load ordering was changed to require explicit declarations of order.  This update tells Drupal that our Theme's style really depends on having FN1 already loaded.