Replace login link with user_login_form
This modified the form and block for the standard user_login_form.  It expects the user_login_form block to be rendered in the footer of all pages.
The form has all the additional links stripped out and the label on the button changed to "Login".

In the Drupal web interface, modify the configuration for the form and block to not render titles.

Updates #269