Updated header and footer
This reworks the homepage header and footer.  The footer is global.  I expect the top, primary links will be replaced with megamenu dropdowns for residents, business, students, etc.

I also added a bit of CSS back in.  The screen.css does not extend from anything.  Rather, it adds additional styling to things that would be unwiedly to do in the markup, using USWDS classes.