Render all regions in page.html.twig
We really want to have all the layout information in a single template.  Drupal expects all node fields to render in only a single region.
To render a sidebar containing asides, calendars, etc., we used to render the sidebar region in preprocess_page, then pass it to the node.
However, it makes the most sense to render all regions in page.html.twig.  Now, the theme creates page variables for all the fields we want to include in the sidebar, instead of rendering them in node.html.twig.